user_service/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
aio_pika
python-jose[cryptography]
passlib[argon2]
email-validator